feat: refactor logging utils to streaminglog package
