+++ cmd/createPromotePipelineJob.go
@@ -114,8 +114,8 @@ var createPromotePipelineJobCmd = &cobra.Command{
 
 		return streaminglog.New(
 			cmd.ErrOrStderr(),
-			getReplicasForJob(apiClient, appName, *jobName),
-			getLogsForJob(apiClient, appName, *jobName),
+			streaminglog.GetReplicasForJob(apiClient, appName, *jobName),
+			streaminglog.GetLogsForJob(apiClient, appName, *jobName),
 			time.Second, // not used
 		).StreamLogs(cmd.Context())
 	},
