+++ cmd/logsEnvironment.go
@@ -15,23 +15,14 @@
 package cmd
 
 import (
-	"context"
 	"errors"
-	"strconv"
-	"strings"
-	"time"
 
-	radixapi "github.com/equinor/radix-cli/generated/radixapi/client"
-	"github.com/equinor/radix-cli/generated/radixapi/client/component"
-	"github.com/equinor/radix-cli/generated/radixapi/client/environment"
 	"github.com/equinor/radix-cli/pkg/client"
 	"github.com/equinor/radix-cli/pkg/config"
 	"github.com/equinor/radix-cli/pkg/flagnames"
 	"github.com/equinor/radix-cli/pkg/settings"
 	"github.com/equinor/radix-cli/pkg/utils/completion"
 	"github.com/equinor/radix-cli/pkg/utils/streaminglog"
-	"github.com/equinor/radix-common/utils/pointers"
-	"github.com/go-openapi/strfmt"
 	"github.com/spf13/cobra"
 )
 
@@ -72,83 +63,11 @@ rx get logs environment --application radix-test --environment dev`,
 
 		return streaminglog.New(
 			cmd.ErrOrStderr(),
-			getComponentReplicasForEnvironment(apiClient, appName, environmentName, previousLog),
-			getComponentLog(apiClient, appName, previousLog),
+			streaminglog.GetComponentReplicasForEnvironment(apiClient, appName, environmentName, previousLog),
+			streaminglog.GetComponentLog(apiClient, appName, previousLog),
 			since,
 		).StreamLogs(cmd.Context())
 	},
-}
-
-type ComponentItem struct {
-	Component string
-	Replica   string
-}
-
-func (c ComponentItem) String() string {
-	return c.Component + "/" + c.Replica
-}
-
-func getComponentLog(apiClient *radixapi.Radixapi, appName string, previous bool) streaminglog.GetLogFunc[ComponentItem] {
-	previousStr := strconv.FormatBool(previous)
-
-	return func(ctx context.Context, item ComponentItem, since time.Time, print func(text string)) error {
-		logParameters := component.NewLogParamsWithContext(ctx)
-		logParameters.WithAppName(appName)
-		logParameters.WithDeploymentName("irrelevant")
-		logParameters.WithComponentName(item.Component)
-		logParameters.WithPodName(item.Replica)
-		logParameters.WithFollow(pointers.Ptr("true"))
-		logParameters.SetSinceTime(pointers.Ptr(strfmt.DateTime(since)))
-		logParameters.WithPrevious(&previousStr)
-
-		resp, err := apiClient.Component.Log(logParameters, nil, streaminglog.CreateLogStreamer(print))
-		if err != nil {
-			return err
-		}
-
-		lines := strings.Split(resp.Payload, "\n")
-		for _, line := range lines {
-			print(line)
-		}
-		print("stream closed.")
-
-		return nil
-	}
-}
-
-// getComponentReplicasForEnvironment returns all replicas for all components in an environment, if previousLog is true, only return replicas once
-func getComponentReplicasForEnvironment(apiClient *radixapi.Radixapi, appName, environmentName string, previousLog bool) streaminglog.GetReplicasFunc[ComponentItem] {
-	return func() ([]ComponentItem, bool, error) {
-		// Get active deployment
-		environmentParams := environment.NewGetEnvironmentParams()
-		environmentParams.SetAppName(appName)
-		environmentParams.SetEnvName(environmentName)
-		environmentDetails, err := apiClient.Environment.GetEnvironment(environmentParams, nil)
-
-		if err != nil {
-			return nil, false, err
-		}
-
-		if environmentDetails == nil || environmentDetails.Payload.ActiveDeployment == nil {
-			return nil, false, errors.New("active deployment was not found in environment")
-		}
-
-		componentReplicas := make([]ComponentItem, 0, 50)
-		for _, component := range environmentDetails.Payload.ActiveDeployment.Components {
-			if component.Name != nil {
-				for _, replica := range component.ReplicaList {
-					componentReplicas = append(componentReplicas, ComponentItem{
-						Component: *component.Name,
-						Replica:   *replica.Name,
-					})
-				}
-				//
-			}
-		}
-
-		// If previous log is requested, return replicas only once
-		return componentReplicas, previousLog, nil
-	}
 }
 
 func init() {
