+++ cmd/logsEnvironmentComponent.go
@@ -17,8 +17,6 @@ package cmd
 import (
 	"errors"
 
-	radixapi "github.com/equinor/radix-cli/generated/radixapi/client"
-	"github.com/equinor/radix-cli/generated/radixapi/client/environment"
 	"github.com/equinor/radix-cli/pkg/client"
 	"github.com/equinor/radix-cli/pkg/config"
 	"github.com/equinor/radix-cli/pkg/flagnames"
@@ -74,45 +72,11 @@ Examples:
 
 		return streaminglog.New(
 			cmd.ErrOrStderr(),
-			getReplicasForComponent(apiClient, appName, environmentName, componentName, previousLog),
-			getComponentLog(apiClient, appName, previousLog),
+			streaminglog.GetReplicasForComponent(apiClient, appName, environmentName, componentName, previousLog),
+			streaminglog.GetComponentLog(apiClient, appName, previousLog),
 			since,
 		).StreamLogs(cmd.Context())
 	},
-}
-
-func getReplicasForComponent(apiClient *radixapi.Radixapi, appName, environmentName, componentName string, previousLog bool) streaminglog.GetReplicasFunc[ComponentItem] {
-	return func() ([]ComponentItem, bool, error) {
-		environmentParams := environment.NewGetEnvironmentParams()
-		environmentParams.SetAppName(appName)
-		environmentParams.SetEnvName(environmentName)
-		environmentDetails, err := apiClient.Environment.GetEnvironment(environmentParams, nil)
-
-		if err != nil {
-			return nil, false, err
-		}
-
-		if environmentDetails == nil || environmentDetails.Payload.ActiveDeployment == nil {
-			return nil, false, errors.New("active deployment was not found in environment")
-		}
-
-		var replicas []ComponentItem
-		for _, comp := range environmentDetails.Payload.ActiveDeployment.Components {
-			if comp.Name != nil && *comp.Name != componentName {
-				continue
-			}
-
-			for _, replica := range comp.ReplicaList {
-				replicas = append(replicas, ComponentItem{
-					Component: *comp.Name,
-					Replica:   *replica.Name,
-				})
-			}
-		}
-
-		// if previous log is requested, return replicas only once
-		return replicas, previousLog, err
-	}
 }
 
 func init() {
